Feat: CRUD로그인 기본 및 SESSION 방문자 카운트 모듈(수정)

+++ backend/nestjs/src/modules/post/post.controller.ts
@@ -1,14 +1,17 @@
 import {
     Controller,
     Get,
-    Post,
+    Post as PostMethod,
     Put,
     Delete,
     Body,
     Param,
     ValidationPipe,
     ParseIntPipe,
+    Req,
+    ForbiddenException,
   } from '@nestjs/common';
+  import { Request } from 'express';
   import { PostService } from './post.service';
   import { CreatePostDto } from './dto/create-post.dto';
   import { UpdatePostDto } from './dto/update-post.dto';
@@ -27,8 +30,11 @@ import {
       return this.postService.findOne(id);
     }
   
-    @Post()
-    create(@Body(ValidationPipe) createPostDto: CreatePostDto) {
+    @PostMethod()
+    create(@Body(ValidationPipe) createPostDto: CreatePostDto, @Req() req:Request) {
+      if (req.session.role !== 0){
+        throw new ForbiddenException('슈퍼관리자만 글쓰기가 가능합니다!');
+      }
       return this.postService.create(createPostDto);
     }
   
@@ -36,12 +42,19 @@ import {
     update(
       @Param('id', ParseIntPipe) id: number,
       @Body(ValidationPipe) updatePostDto: UpdatePostDto,
+      @Req() req: Request,
     ) {
+      if(req.session.role !== 0){
+        throw new ForbiddenException('슈퍼관리자만 수정이 가능합니다!');
+      }
       return this.postService.update(id, updatePostDto);
     }
   
     @Delete(':id')
-    remove(@Param('id', ParseIntPipe) id: number) {
+    remove(@Param('id', ParseIntPipe) id: number, @Req() req:Request) {
+      if(req.session.role !==0){
+        throw new ForbiddenException('슈퍼관리자만 삭제가 가능합니다!');
+      }
       return this.postService.remove(id);
     }
   }
